add dune dex/lending users, remove stale entries
